flatten the account state root in one single tree

+++ executor/src/error.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub enum Error {
     ContractCall(u8),
-    MissingAccount(u32),
+    MissingAccount(u64),
     /// balance, required_amount
     BalanceNotEnough(u64, u64),
     /// expected nonce, tx's nonce
